BE/requirements.txt
# BE/requirements.txt

Flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1
psutil==6.0.0

# Drivers
pymongo==4.8.0
pymysql==1.1.1
psycopg2-binary==2.9.9
oracledb==2.3.0

# (옵션) 테스트
pytest==8.3.3
